htmltopdf/requirements.txt
Django>=5.0
djangorestframework
weasyprint
gunicorn
xhtml2pdf
boto3
django-storages
django-minio-storage